Cleaning up excessive binaries
Improving python tabbing > spacing
Extending jet collections in ntuples from 4 to 10

+++ SUSYBSMAnalysis/python/DileptonPreselector_cfi.py
@@ -2,7 +2,7 @@ import FWCore.ParameterSet.Config as cms
 
 dileptonPreseletor = cms.EDFilter("DileptonPreselector",
 	muons = cms.InputTag("slimmedMuons"),
-	nMuons = cms.double(2),
+	nMuons = cms.double(1),
 	ptCut = cms.double(40),	
 
 )
